requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
pycparser==2.19
pycricbuzz==2.4
python-telegram-bot==12.0.0b1
requests==2.21.0
schedule==0.6.0
six==1.12.0
tornado==6.0.2
urllib3==1.24.1
